pyproject.toml
[tool.poetry]
name = "media-escolar"
version = "0.1.0"
description = ""
authors = ["LuizFernando110 <luizloona@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
